+++ app/src/main/java/org/simple/clinic/registerorlogin/AuthenticationActivity.kt
@@ -145,6 +145,6 @@ class AuthenticationActivity : AppCompatActivity(), AuthenticationUiActions {
   }
 
   override fun openRegistrationPhoneScreen() {
-    router.clearHistoryAndPush(RegistrationPhoneScreenKey().wrap())
+    router.clearHistoryAndPush(RegistrationPhoneScreenKey())
   }
 }